move integration tests for create to its own file
