+++ test/integration/commands/test_autosubmit_commands.py
@@ -17,36 +17,6 @@
 
 import pytest
 
-from autosubmit.config.basicconfig import BasicConfig
-from autosubmit.history.database_managers.experiment_history_db_manager import (
-    SqlAlchemyExperimentHistoryDbManager,
-)
-
-
-@pytest.mark.parametrize("noplot", [True, False])
-def test_create_noplot_calls_generate_output(as_exp, mocker, noplot):
-    """Test that create calls generate_output when noplot is False and does not call it when noplot is True."""
-    db_manager = SqlAlchemyExperimentHistoryDbManager(
-        as_exp.expid, BasicConfig.JOBDATA_DIR, f"job_data_{as_exp.expid}.db"
-    )
-
-    db_manager.initialize()
-
-    mock_generate_output = mocker.patch(
-        "autosubmit.monitor.monitor.Monitor.generate_output"
-    )
-
-    as_exp.autosubmit.create(
-        as_exp.expid,
-        noplot=noplot,
-        hide=True,
-    )
-
-    if noplot:
-        mock_generate_output.assert_not_called()
-    else:
-        mock_generate_output.assert_called_once()
-
 
 @pytest.mark.parametrize("noclean,uncompress", [
     (True, True),
@@ -70,57 +40,3 @@ def test_archive_noncreated_experiment(as_exp):
 
 def test_unarchive_nonarchived_experiment(as_exp):
     assert not as_exp.autosubmit.unarchive(as_exp.expid)
-
-
-def test_create_cw_calls_generate_scripts_andor_wrappers_without_plt(
-    autosubmit_exp, general_data, mocker
-):
-    """create -cw must call generate_scripts_andor_wrappers even without -plt.
-
-    Before the fix the wrapper generation was nested inside ``if not noplot:``
-    so ``create -cw`` without ``-plt`` silently skipped preview-table
-    population.
-    """
-    mock_gen = mocker.patch(
-        "autosubmit.autosubmit.Autosubmit.generate_scripts_andor_wrappers"
-    )
-
-    exp_data = {
-        "EXPERIMENT": {
-            "DATELIST": "20200101",
-            "MEMBERS": "fc0",
-            "CHUNKSIZEUNIT": "month",
-            "CHUNKSIZE": 1,
-            "NUMCHUNKS": 1,
-            "CALENDAR": "standard",
-        },
-        "JOBS": {
-            "SLURMJOB": {
-                "SCRIPT": "sleep 0",
-                "RUNNING": "chunk",
-                "WALLCLOCK": "02:00",
-                "PLATFORM": "TEST_SLURM",
-            },
-        },
-        "WRAPPERS": {
-            "SIMPLE_WRAPPER": {
-                "TYPE": "vertical",
-                "JOBS_IN_WRAPPER": "SLURMJOB",
-                "MAX_WRAPPED_V": 2,
-                "MIN_WRAPPED_V": 2,
-            },
-        },
-    }
-    config_data = general_data | exp_data
-
-    exp = autosubmit_exp(experiment_data=config_data, include_jobs=False, create=True)
-
-    exp.autosubmit.create(
-        exp.expid,
-        noplot=True,
-        hide=True,
-        check_wrappers=True,
-    )
-
-    mock_gen.assert_called_once()
-
